refactor(client): moving git logic to hook
